tests and bugfixes
