+++ src/neuroglancer_annotation_ui/statebuilder/__init__.py
@@ -1,1 +1,1 @@
-from .statebuilder import StateBuilder, FilteredDataStateBuilder, ChainedStateBuilder, build_state
+from .statebuilder import StateBuilder, FilteredDataStateBuilder, ChainedStateBuilder, build_state_direct
