+++ src/neuroglancer_annotation_ui/statebuilder/statebuilder.py
@@ -1,16 +1,16 @@
 from neuroglancer_annotation_ui import EasyViewer, annotation
-from annotationframeworkclient.infoservice import InfoServiceClient
 import pandas as pd
 import numpy as np
-import warnings
+from IPython.display import HTML
 from collections import defaultdict
-from .utils import bucket_of_values
+from .utils import bucket_of_values, make_basic_dataframe, sources_from_infoclient
 
 
-def build_state_flat(selected_ids=[], point_annotations={},
-                     line_annotations={}, sphere_annotations={},
-                     return_as='url', render_kwargs={},
-                     state_kwargs={}):
+def build_state_direct(dataset_name=None, selected_ids=[], point_annotations={},
+                      line_annotations={}, sphere_annotations={},
+                      default_anno_layer_name='annos',
+                      return_as='url', render_kws={},
+                      state_kws={}):
     """Build a Neuroglancer state from data directly, using the statebuilder as an intermediate layer.
 
     Parameters
@@ -54,88 +54,9 @@ def build_state_flat(selected_ids=[], point_annotations={},
         if not isinstance(sphere_annotations, dict):
             sphere_annotations = {default_anno_layer_name: sphere_annotations}
     
-    df, pals, lals, sals = _make_basic_dataframe(point_annotations, line_annotations, sphere_annotations)
-    sb = StateBuilder(point_annotations=pals, line_annotations=lals, sphere_annotations=sals, **kwargs)
-    return sb.render_state(data=df, return_as='url', **render_kwargs)
-
-def _make_basic_dataframe(point_annotations={}, line_annotations={}, sphere_annotations={}):
-    """Makes a dataframe containing pre-computed data to be read by a StateBuilder object.
-    """
-    dfs = []
-
-    pals = {}
-    for ii, (ln, data) in enumerate(point_annotations.items()):
-        col_name = f'point_{ii}'
-        if isinstance(data, np.ndarray):
-            data = data.tolist()
-        df=pd.DataFrame()
-        df[col_name] = data
-        dfs.append(df)
-        pals[ln] = [col_name]
-    
-    lals = {}
-    for ii, (ln, (data_a, data_b)) in enuemrate(line_annotations.items()):
-        col_name_a = f'line_{ii}_a'
-        col_name_b = f'line_{ii}_b'
-        if isinstance(data_a, np.ndarray):
-            data_a = data_a.tolist()
-        if isinstance(data_b, np.ndarray):
-            data_b = data_b.tolist()
-        df = pd.DataFrame()
-        df[col_name_a] = data_a
-        df[col_name_b] = data_b
-        dfs.append(df)
-        lals[ln] = [[col_name_a, col_name_b]]
-
-    sals = {}
-    for ii, (ln, (data_c, data_r)) in enuemrate(sphere_annotations.items()):
-        col_name_c = f'sphere_{ii}_c'
-        col_name_r = f'sphere_{ii}_r'
-        if isinstance(data_c, np.ndarray):
-            data_c = data_c.tolist()
-        if isinstance(data_r, np.ndarray):
-            data_r = data_r.tolist()
-        df = pd.DataFrame()
-        df[col_name_c] = data_c
-        df[col_name_r] = data_r
-        dfs.append(df)
-        sals[ln] = [[col_name_c, col_name_r]]
-    
-    df_out = pd.concat(dfs, sort=False)
-    return df_out, pals, lals, sals
-
-
-def sources_from_infoclient(dataset_name, segmentation_type='default', image_layer_name='img', seg_layer_name='seg'):
-    """Generate an img_source and seg_source dict from the info service. Will default to graphene and fall back to flat segmentation, unless otherwise specified. 
-    
-    Parameters
-    ----------
-    dataset_name : str
-        InfoService dataset name
-    segmentation_type : 'default', 'graphene' or 'flat', optional
-        Choose which type of segmentation to use. 'default' will try graphene first and fall back to flat. Graphene or flat will
-        only use the specified type or give nothing. By default 'default'. 
-    image_layer_name : str, optional
-        Layer name for the imagery, by default 'img'
-    seg_layer_name : str, optional
-        Layer name for the segmentation, by default 'seg'
-    """
-    info_client = InfoServiceClient(dataset_name=dataset_name)
-    image_source = {image_layer_name: info_client.image_source(format_for='neuroglancer')}
-    
-    if segmentation_type == 'default':
-        if info_client.pychunkedgraph_segmentation_source() is None:
-            segmentation_type = 'flat'
-        else:
-            segmentation_type = 'graphene'
-
-    if segmentation_type == 'graphene':
-        seg_source = {seg_layer_name : info_client.pychunkedgraph_segmentation_source(format_for='neuroglancer')}
-    elif segmentation_type == 'flat':
-        seg_source = {seg_layer_name: info_client.flat_segmentation_source(format_for='neuroglancer')}
-    else:
-        seg_source = {}
-    return image_source, seg_source
+    df, pals, lals, sals = make_basic_dataframe(point_annotations, line_annotations, sphere_annotations)
+    sb = StateBuilder(dataset_name=dataset_name, point_annotations=pals, line_annotations=lals, sphere_annotations=sals, **state_kws)
+    return sb.render_state(data=df, return_as=return_as, **render_kws)
 
 
 class StateBuilder():
@@ -368,13 +289,9 @@ class StateBuilder():
         elif return_as == 'url':
             out = self._temp_viewer.as_url(prefix=url_prefix)
             self.initialize_state()
-            # if len(out) > MAX_URL_LENGTH:
-            #     warnings.warn('URL exceeds max length for many web browsers', Warning)
             return out
         elif return_as == 'html':
-            out = self._temp_viewer.as_url(prefix=url_prefix, as_html=True, link_text=link_text)
-            # if len(out) > MAX_URL_LENGTH + len(link_text) + 31: # 31=Length of html characters in the as_url function
-            #     warnings.warn('URL exceeds max length for many web browsers', Warning)
+            out = HTML(self._temp_viewer.as_url(prefix=url_prefix, as_html=True, link_text=link_text))
             self.initialize_state()
             return out
         elif return_as == 'json':
